Set index (#36)
implement distributed set_index.
