+++ sparsity/test/test_sparse_frame.py
@@ -56,75 +56,6 @@ def mock_s3_fs(bucket, data=None):
         m.stop()
 
 
-# 2017 starts with a sunday
-@pytest.fixture()
-def sampledata():
-    def gendata(n, categorical=False):
-        sample_data = pd.DataFrame(
-            dict(date=pd.date_range("2017-01-01", periods=n)))
-        sample_data["weekday"] = sample_data.date.dt.weekday_name
-        sample_data["weekday_abbr"] = sample_data.weekday.apply(
-            lambda x: x[:3])
-
-        if categorical:
-            sample_data['weekday'] = sample_data['weekday'].astype('category')
-            sample_data['weekday_abbr'] = sample_data['weekday_abbr'] \
-                .astype('category')
-
-        sample_data["id"] = np.tile(np.arange(7), len(sample_data) // 7 + 1)[
-                            :len(sample_data)]
-        return sample_data
-
-    return gendata
-
-
-@pytest.fixture()
-def sample_frame_labels():
-    return SparseFrame(np.identity(5),
-                       columns = list('ABCDE'),
-                       index = list('VWXYZ'))
-
-@pytest.fixture()
-def weekdays():
-    return ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday',
-            'Friday', 'Saturday']
-
-
-@pytest.fixture()
-def weekdays_abbr(weekdays):
-    return list(map(lambda x: x[:3], weekdays))
-
-
-@pytest.fixture()
-def groupby_frame():
-    shuffle_idx = np.random.permutation(np.arange(100))
-    index = np.tile(np.arange(10), 10)
-    data = np.vstack([np.identity(10) for _ in range(10)])
-    t = SparseFrame(data[shuffle_idx, :], index=index[shuffle_idx])
-    return t
-
-
-@pytest.fixture()
-def sf_midx():
-    midx = pd.MultiIndex.from_arrays(
-        [pd.date_range("2016-10-01", periods=5),
-         np.arange(5)]
-    )
-    cols = list('ABCDE')
-    sf = SparseFrame(np.identity(5), index=midx, columns=cols)
-    return sf
-
-@pytest.fixture()
-def sf_midx_int():
-    midx = pd.MultiIndex.from_arrays(
-        [np.concatenate([np.ones(4), np.zeros(1)]),
-         pd.date_range("2016-10-01", periods=5)]
-    )
-    cols = list('ABCDE')
-    sf = SparseFrame(np.identity(5), index=midx, columns=cols)
-    return sf
-
-
 def test_empty_init():
     sf = SparseFrame(np.array([]), index=[], columns=['A', 'B'])
     assert sf.data.shape == (0, 2)
